modified jni.cc and add java-api-examples
