+++ java-api-examples/.gitattributes
