+++ java-api-examples/app/modelconfig.cfg
@@ -1,5 +1,5 @@
 #jni-java .so lib root dir
-solibpath=/sherpa/forgithub/sherpa-onnx/build/lib/libsherpa-onnx-jni-java.so  
+solibpath=/sherpa/forgithub/sherpa-onnx/build/lib/libsherpa-onnx-jni.so  
 sample_rate=16000                  
 feature_dim=80
 rule1_min_trailing_silence=2.4
