+++ java-api-examples/app/src/main/java/com/k2fsa/sherpaonnx/FeatureConfig.java
