+++ java-api-examples/settings.gradle
